requirements.txt
#fastapi>=0.110
#uvicorn[standard]>=0.27
#edge-tts>=6.1.12
fastapi
uvicorn[standard]
edge-tts
ruff
pre-commit	
google-cloud-storage
